requirements.txt
python-telegram-bot==20.7
tonsdk==1.0.9
aiohttp==3.9.3